GVT-3548: Also separate operational point row into its own file
