+++ ui/src/tool-panel/location-track/location-track-operational-point-links-infobox.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import Infobox from 'tool-panel/infobox/infobox';
-import { LayoutLocationTrack, OperationalPoint } from 'track-layout/track-layout-model';
+import { LayoutLocationTrack } from 'track-layout/track-layout-model';
 import { useLocationTrackInfoboxExtras } from 'track-layout/track-layout-react-utils';
-import { LayoutContext, TrackMeter } from 'common/common-model';
-import { Button, ButtonSize, ButtonVariant } from 'vayla-design-lib/button/button';
+import { LayoutContext } from 'common/common-model';
 import styles from './location-track-operational-point-links-infobox.scss';
 import { ChangeTimes } from 'common/common-slice';
-import NavigableTrackMeter from 'geoviite-design-lib/track-meter/navigable-track-meter';
-import { OperationalPointBadge } from 'geoviite-design-lib/operational-point/operational-point-badge';
 import { OnSelectOptions } from 'selection/selection-model';
 import { ShowMoreButton } from 'show-more-button/show-more-button';
 import {
@@ -18,9 +15,7 @@ import {
 import { LoaderStatus, useLoaderWithStatus } from 'utils/react-utils';
 import { getManyOperationalPoints } from 'track-layout/layout-operational-point-api';
 import InfoboxContent from 'tool-panel/infobox/infobox-content';
-import infoboxStyles from 'tool-panel/infobox/infobox.module.scss';
-import { createClassName } from 'vayla-design-lib/utils';
-import { LocationTrackDetachOperationalPointDialog } from './dialog/location-track-detach-operational-point-dialog';
+import { LocationTrackOperationalPointRow } from './location-track-operational-point-row';
 
 const maxOperationalPointsToDisplay = 10;
 
@@ -132,77 +127,3 @@ export const LocationTrackOperationalPointLinksInfobox: React.FC<
     );
 };
 
-type LocationTrackOperationalPointRowProps = {
-    operationalPoint: OperationalPoint;
-    address: TrackMeter | undefined;
-    layoutContext: LayoutContext;
-    locationTrack: LayoutLocationTrack;
-    onSelect: (items: OnSelectOptions) => void;
-};
-
-const LocationTrackOperationalPointRow: React.FC<LocationTrackOperationalPointRowProps> = ({
-    operationalPoint,
-    address,
-    layoutContext,
-    locationTrack,
-    onSelect,
-}) => {
-    const { t } = useTranslation();
-    const [showDetachDialog, setShowDetachDialog] = React.useState(false);
-    const remarkClassNames = createClassName(
-        styles['location-track-operational-point-links-infobox-list__remark'],
-        infoboxStyles['infobox__list-cell--strong'],
-    );
-
-    return (
-        <>
-            <div>
-                <OperationalPointBadge
-                    operationalPoint={operationalPoint}
-                    onClick={() =>
-                        onSelect({
-                            operationalPoints: [operationalPoint.id],
-                            selectedTab: { id: operationalPoint.id, type: 'OPERATIONAL_POINT' },
-                        })
-                    }
-                />
-            </div>
-            <div className={infoboxStyles['infobox__list-cell--strong']}>
-                {!address ? (
-                    t('tool-panel.location-track.operational-point-links.no-location')
-                ) : (
-                    <NavigableTrackMeter
-                        trackMeter={address}
-                        displayDecimals={false}
-                        location={operationalPoint.location}
-                    />
-                )}
-            </div>
-            <div className={remarkClassNames}>
-                {operationalPoint.state === 'DELETED' &&
-                    t('tool-panel.location-track.operational-point-links.not-existing')}
-            </div>
-            <div>
-                {layoutContext.publicationState === 'DRAFT' && (
-                    <Button
-                        size={ButtonSize.SMALL}
-                        variant={ButtonVariant.GHOST}
-                        onClick={() => setShowDetachDialog(true)}>
-                        {t('tool-panel.location-track.operational-point-links.detach')}
-                    </Button>
-                )}
-            </div>
-            {showDetachDialog && (
-                <LocationTrackDetachOperationalPointDialog
-                    layoutContext={layoutContext}
-                    locationTrackId={locationTrack.id}
-                    locationTrackName={locationTrack.name}
-                    operationalPointId={operationalPoint.id}
-                    operationalPointName={operationalPoint.name}
-                    onClose={() => setShowDetachDialog(false)}
-                    onDetached={() => setShowDetachDialog(false)}
-                />
-            )}
-        </>
-    );
-};
